agents/autonomous-agents-system/skill-matcher-agent/requirements.txt
uagents>=0.4.0
hyperon-experimental>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
